Allow to use "finalName" as destFileName default
Maven allows to set a "finalName" - the name which will be used when a
project builds jars or other files in the /target folder.
This commit allows to set a <useFinalName/> parameter which will
load the POM of a dependency and use the finalName field from it
instead of the default "artifactId-version-classifier" scheme.